Move classes to reference layout
