+++ dynamodb/src/main/scala/akka/stream/alpakka/dynamodb/AwsClientSettings.scala
@@ -2,11 +2,11 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.dynamodb.impl
+package akka.stream.alpakka.dynamodb
 
 import com.amazonaws.auth.AWSCredentialsProvider
 
-abstract class ClientSettings {
+abstract class AwsClientSettings {
   val region: String
   val host: String
   val port: Int
