+++ dynamodb/src/main/scala/akka/stream/alpakka/dynamodb/impl/DynamoClientImpl.scala
@@ -5,18 +5,24 @@
 package akka.stream.alpakka.dynamodb.impl
 
 import akka.actor.ActorSystem
+import akka.annotation.InternalApi
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.MediaType.NotCompressible
 import akka.http.scaladsl.model.{ContentType, MediaType}
 import akka.http.scaladsl.settings.ConnectionPoolSettings
 import akka.stream.Materializer
+import akka.stream.alpakka.dynamodb.DynamoSettings
 import akka.stream.alpakka.dynamodb.impl.AwsClient.{AwsConnect, AwsRequestMetadata}
 import com.amazonaws.AmazonServiceException
 import com.amazonaws.http.HttpResponseHandler
 
 import scala.concurrent.ExecutionContextExecutor
 
-class DynamoClientImpl(
+/**
+ * INTERNAL API
+ */
+@InternalApi
+private[dynamodb] class DynamoClientImpl(
     val settings: DynamoSettings,
     val errorResponseHandler: HttpResponseHandler[AmazonServiceException]
 )(implicit protected val system: ActorSystem, implicit protected val materializer: Materializer)
